Random color, H with floats, minor changes

+++ ui/main/modeling/linear/step6.py
@@ -64,7 +64,7 @@ class Step6(QWidget):
 
         value = item.text()
         try:
-            int(value)
+            float(value)
             if not item.valid:
                 item.valid = True
                 self.errors -= 1
